[*] 优化代码，修改升级版本后适配问题

+++ ingot-service/ingot-pms/ingot-pms-provider/src/main/java/com/ingot/cloud/pms/service/domain/impl/SysDeptServiceImpl.java
@@ -68,7 +68,7 @@ public class SysDeptServiceImpl extends BaseServiceImpl<SysDeptMapper, SysDept> 
         assertI18nService.checkOperation(success, "SysDeptServiceImpl.ExistLeaf");
 
         // 取消关联角色
-        int count = sysRoleDeptService.count(Wrappers.<SysRoleDept>lambdaQuery()
+        long count = sysRoleDeptService.count(Wrappers.<SysRoleDept>lambdaQuery()
                 .eq(SysRoleDept::getDeptId, id));
         if (count > 0) {
             success = sysRoleDeptService.remove(Wrappers.<SysRoleDept>lambdaQuery()
